backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
pydantic==2.9.2
websockets==13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-genai==0.2.2
celery==5.4.0
redis==5.0.3
python-owasp-zap-v2.4==0.0.22
slowapi==0.1.9
